reduce the shift size of the address down to the lower half to improve performance.

+++ projects/common/deception/include/Deception.h
@@ -301,7 +301,7 @@ namespace Deception {
             };
             using Set_t = CacheSet;
             static constexpr uint8_t computeIndex(Address input) noexcept {
-                return (input >> Line_t::ShiftAmount) & LineMask;
+                return (static_cast<uint16_t>(input) >> Line_t::ShiftAmount) & LineMask;
             }
             void clear() noexcept {
                 for (auto set : sets) {
